+++ src/main/java/ajaajaja/debuging_rounge/domain/question/repository/QuestionRepository.java
@@ -1,6 +1,6 @@
 package ajaajaja.debuging_rounge.domain.question.repository;
 
-import ajaajaja.debuging_rounge.domain.question.domain.Question;
+import ajaajaja.debuging_rounge.domain.question.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
